test: 코드 위치 수정
